Sharing lots of love today
Moved types to messages/index.ts, as it makes more sense there
